requirements.txt
asttokens==2.2.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
tweepy==4.12.1
twitter==2.0a2
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
